+++ packages/atomic/src/components/common/facets/timeframe-facet-common.tsx
@@ -197,7 +197,7 @@ export class TimeframeFacetCommon {
 
   private registerFacetToStore() {
     const facetInfo: FacetInfo = {
-      label: this.props.label,
+      label: () => this.props.bindings.i18n.t(this.props.label),
       facetId: this.facetId!,
       element: this.props.host,
     };