+++ packages/atomic/src/components/common/result-list/result-list-common-interface.ts
@@ -9,6 +9,7 @@ import {FocusTargetController} from '../../../utils/accessibility-utils';
 import {ResultTemplateProvider} from './result-template-provider';
 import {AtomicCommonStore, AtomicCommonStoreData} from '../interface/store';
 import {VNode} from '@stencil/core';
+import {InteractiveResult} from '@coveo/headless';
 
 export interface ResultListRenderer {
   getResultId(result: AnyResult): string;
@@ -24,13 +25,15 @@ export interface ResultListCommonProps<
   loadingFlag: string;
   resultTemplateProvider: ResultTemplateProvider;
   nextNewResultTarget: FocusTargetController;
-  getDisplay(): ResultDisplayLayout;
+  getLayoutDisplay(): ResultDisplayLayout;
+  getResultDisplay(): ResultDisplayLayout;
   getDensity(): ResultDisplayDensity;
   getImageSize(): ResultDisplayImageSize;
   getResultListState(): ResultListCommonState<SpecificResult>;
   getNumberOfPlaceholders(): number;
   getResultRenderingFunction(): ResultRenderingFunction;
   renderResult(props: ResultRendererProps<SpecificResult>): VNode;
+  getInteractiveResult(result: AnyResult): InteractiveResult;
 }
 
 export interface ResultListCommonState<SpecificResult extends AnyResult> {
@@ -52,6 +55,7 @@ export interface ResultRendererProps<
   key?: string;
   part?: string;
   result: SpecificResult;
+  interactiveResult: InteractiveResult;
   content?: ParentNode;
   loadingFlag?: string;
   store: AtomicCommonStore<AtomicCommonStoreData>;