+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.32.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.32.0...@coveo/atomic-angular@1.32.1) (2022-10-11)
+
+
+### Bug Fixes
+
+* **atomic:** modify result-localized-text key property ([#2493](https://github.com/coveo/ui-kit/issues/2493)) ([a6eb820](https://github.com/coveo/ui-kit/commit/a6eb82036547527507e8a90c1e69b89797daf18e))
+
+
+
+
+
+# [1.32.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.31.0...@coveo/atomic-angular@1.32.0) (2022-10-07)
+
+
+### Features
+
+* **atomic:** add calendar time support + result-timespan component ([#2475](https://github.com/coveo/ui-kit/issues/2475)) ([59ff2a8](https://github.com/coveo/ui-kit/commit/59ff2a8ba0b52fd1afd9af0ba50521c42cdd1c76))
+
+
+
+
+
+# [1.31.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.18...@coveo/atomic-angular@1.31.0) (2022-10-06)
+
+
+### Features
+
+* **atomic,headless:** add rec result & template components, add missing rec state, add recommendations list styling & layout ([#2484](https://github.com/coveo/ui-kit/issues/2484)) ([c3fb9eb](https://github.com/coveo/ui-kit/commit/c3fb9ebab3181da7b55a3c8152822455399fbfa1))
+* **atomic:** add atomic-result-localized-text component ([#2483](https://github.com/coveo/ui-kit/issues/2483)) ([662dc14](https://github.com/coveo/ui-kit/commit/662dc14e1164e62ed2ae3c8356c610f90d0b8969))
+
+
+
+
+
+## [1.30.18](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.17...@coveo/atomic-angular@1.30.18) (2022-10-05)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
 ## [1.30.17](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.16...@coveo/atomic-angular@1.30.17) (2022-10-03)
 
 **Note:** Version bump only for package @coveo/atomic-angular