+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.108.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.107.0...@coveo/headless@1.108.0) (2022-10-11)
+
+
+### Features
+
+* **atomic,headless:** use different InteractiveResult per engine/interface ([#2494](https://github.com/coveo/ui-kit/issues/2494)) ([ce4ab2b](https://github.com/coveo/ui-kit/commit/ce4ab2ba182ecaecda8a0f4e98d208827859fed4))
+
+
+
+
+
+# [1.107.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.106.0...@coveo/headless@1.107.0) (2022-10-06)
+
+
+### Features
+
+* **atomic,headless:** add rec result & template components, add missing rec state, add recommendations list styling & layout ([#2484](https://github.com/coveo/ui-kit/issues/2484)) ([c3fb9eb](https://github.com/coveo/ui-kit/commit/c3fb9ebab3181da7b55a3c8152822455399fbfa1))
+
+
+
+
+
+# [1.106.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.105.2...@coveo/headless@1.106.0) (2022-10-05)
+
+
+### Features
+
+* **atomic:** fields to include refactor ([#2480](https://github.com/coveo/ui-kit/issues/2480)) ([c86cd92](https://github.com/coveo/ui-kit/commit/c86cd92d6b93feb3ab6f10e10d3ecb9fbe3c64e9))
+
+
+
+
+
 ## [1.105.2](https://github.com/coveo/ui-kit/compare/@coveo/headless@1.105.1...@coveo/headless@1.105.2) (2022-09-22)